Move ContextTagsEventProcessor from spring boot to spring module
