+++ sentry-spring-boot-starter/api/sentry-spring-boot-starter.api
@@ -3,11 +3,6 @@ public final class io/sentry/spring/boot/BuildConfig {
 	public static final field VERSION_NAME Ljava/lang/String;
 }
 
-public final class io/sentry/spring/boot/ContextTagsEventProcessor : io/sentry/EventProcessor {
-	public fun <init> (Lio/sentry/SentryOptions;)V
-	public fun process (Lio/sentry/SentryEvent;Ljava/util/Map;)Lio/sentry/SentryEvent;
-}
-
 public class io/sentry/spring/boot/InAppIncludesResolver : org/springframework/context/ApplicationContextAware {
 	public fun <init> ()V
 	public fun resolveInAppIncludes ()Ljava/util/List;
