+++ sentry-spring/src/main/java/io/sentry/spring/ContextTagsEventProcessor.java
@@ -1,4 +1,4 @@
-package io.sentry.spring.boot;
+package io.sentry.spring;
 
 import io.sentry.EventProcessor;
 import io.sentry.SentryEvent;
