+++ sentry-spring/src/test/kotlin/io/sentry/spring/ContextTagsEventProcessorTest.kt
@@ -1,4 +1,4 @@
-package io.sentry.spring.boot
+package io.sentry.spring
 
 import io.sentry.SentryEvent
 import io.sentry.SentryOptions
